backend/requirements.txt
amqp==2.6.1
appdirs==1.4.3
asgiref==3.2.7
astroid==2.3.3
attrs==19.3.0
backcall==0.1.0
bandit==1.6.2
billiard==3.6.3.0
black==19.10b0
boto==2.49.0
brotlipy==0.7.0
celery==4.4.7
certifi==2020.4.5.1
cffi==1.14.0
cfgv==3.1.0
chardet==3.0.4
click==7.1.1
colorama==0.4.3
coverage==5.0.4
cryptography==2.9
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.0
dj-database-url==0.5.0
Django==2.2.15
django-cors-headers==3.4.0
django-crontab==0.7.1
django-import-export==2.3.0
django-js-reverse==0.9.1
django-log-request-id==1.4.1
django-model-utils==4.0.0
django-naomi==0.8
django-rest-auth==0.9.5
django-ses==0.8.14
django-tables2==2.3.1
django-webpack-loader==0.7.0
djangorestframework==3.11.0
dodgy==0.2.1
dparse==0.5.0
et-xmlfile==1.0.1
fcm-django==0.3.4
filelock==3.0.12
future==0.18.2
gitdb==4.0.2
GitPython==3.1.0
gunicorn==20.0.4
identify==1.4.14
idna==2.9
importlib-metadata==1.6.0
importlib-resources==1.4.0
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
jdcal==1.4.1
jedi==0.16.0
kombu==4.6.11
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
model-mommy==2.0.0
nodeenv==1.3.5
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.3
packaging==20.3
parso==0.6.2
pathspec==0.7.0
pbr==5.4.5
pep8-naming==0.4.1
pickleshare==0.7.5
pipenv==2018.11.26
pre-commit==2.2.0
prompt-toolkit==3.0.5
prospector==1.2.0
psutil==5.7.0
psycopg2==2.8.5
pycodestyle==2.4.0
pycparser==2.20
pydocstyle==5.0.2
pyfcm==1.4.7
pyflakes==2.1.1
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.4.4
pylint-celery==0.3
pylint-django==2.0.12
pylint-flask==0.6
pylint-plugin-utils==0.6
pyparsing==2.4.7
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3.1
redis==3.4.1
regex==2020.4.4
requests==2.23.0
requests-oauthlib==1.3.0
requirements-detector==0.6
safety==1.8.7
sentry-sdk==0.14.3
setoptconf==0.2.0
six==1.14.0
smmap==3.0.1
snowballstemmer==2.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
stevedore==1.32.0
tablib==1.1.0
toml==0.10.0
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.8
vine==1.3.0
virtualenv==20.0.16
virtualenv-clone==0.5.4
wcwidth==0.1.9
whitenoise==5.0.1
wrapt==1.11.2
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
